Add files via upload
The Sparks Foundation - Payment Gateway Integration(task-1) .
A simple website where payment gateway is integrated. 
There is a simple donate button on homepage. 
On clicking the donate button, the user will land on the payment page where user can select the amount to be paid and the payment type, e.g.
credit card, Paypal, etc.
Once the payment is done an invoice will be generated and email will be sent to the user for the payment received. 
The invoice will contain the amount.